# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/documents": {
      "post": {
        "operationId": "DocumentsController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilgrimDocumentShape"
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DocumentsController_listByBooking",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PilgrimDocumentShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/documents/mine": {
      "get": {
        "operationId": "DocumentsController_listMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PilgrimDocumentShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/documents/expiry-alerts": {
      "get": {
        "operationId": "DocumentsController_getExpiryAlerts",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentExpiryAlertShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/documents/{id}/access-url": {
      "get": {
        "operationId": "DocumentsController_getAccessUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/documents/{id}/validate": {
      "patch": {
        "operationId": "DocumentsController_validate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilgrimDocumentShape"
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/documents/{id}/reject": {
      "patch": {
        "operationId": "DocumentsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectDocumentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PilgrimDocumentShape"
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/disputes": {
      "post": {
        "operationId": "DisputesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDisputeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeShape"
                }
              }
            }
          }
        },
        "tags": [
          "disputes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "DisputesController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "open",
                "agency_responded",
                "escalated",
                "resolved",
                "closed"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DisputeShape"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "disputes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/disputes/{id}": {
      "get": {
        "operationId": "DisputesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeShape"
                }
              }
            }
          }
        },
        "tags": [
          "disputes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/disputes/{id}/messages": {
      "post": {
        "operationId": "DisputesController_addMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeShape"
                }
              }
            }
          }
        },
        "tags": [
          "disputes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/disputes/{id}/resolve": {
      "post": {
        "operationId": "DisputesController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeShape"
                }
              }
            }
          }
        },
        "tags": [
          "disputes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/disputes/{id}/escalate": {
      "post": {
        "operationId": "DisputesController_escalate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeShape"
                }
              }
            }
          }
        },
        "tags": [
          "disputes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/disputes/{id}/decision": {
      "post": {
        "operationId": "DisputesController_decide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeDecisionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisputeShape"
                }
              }
            }
          }
        },
        "tags": [
          "disputes"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Plateforme Oumra & Hadj — API",
    "description": "API centrale : pèlerins, agences, forfaits, réservations, paiements, documents, rites, groupes, notifications.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UploadDocumentDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string"
          },
          "type": {
            "enum": [
              "passport",
              "visa",
              "flight_ticket",
              "vaccination_certificate"
            ],
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "bookingId",
          "type"
        ]
      },
      "PilgrimDocumentShape": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bookingId": {
            "type": "string"
          },
          "pilgrimId": {
            "type": "string"
          },
          "type": {
            "enum": [
              "passport",
              "visa",
              "flight_ticket",
              "vaccination_certificate"
            ],
            "type": "string"
          },
          "storageRef": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "validated",
              "rejected"
            ],
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "bookingId",
          "pilgrimId",
          "type",
          "storageRef",
          "status"
        ]
      },
      "DocumentExpiryAlertShape": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "passport",
              "visa",
              "flight_ticket",
              "vaccination_certificate"
            ]
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "type",
          "expiresAt",
          "status"
        ]
      },
      "RejectDocumentDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 3
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "CreateDisputeDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "payment",
              "refund",
              "accommodation",
              "transport",
              "documents",
              "service",
              "other"
            ]
          },
          "subject": {
            "type": "string",
            "minLength": 5,
            "maxLength": 140
          },
          "message": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000
          }
        },
        "required": [
          "bookingId",
          "category",
          "subject",
          "message"
        ]
      },
      "DisputeMessageShape": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "authorRole": {
            "type": "string",
            "enum": [
              "pilgrim",
              "agency",
              "guide",
              "admin"
            ]
          },
          "authorName": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "authorRole",
          "authorName",
          "content",
          "createdAt"
        ]
      },
      "DisputeShape": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "bookingId": {
            "type": "string"
          },
          "packageTitle": {
            "type": "string"
          },
          "agencyId": {
            "type": "string"
          },
          "agencyName": {
            "type": "string"
          },
          "pilgrimName": {
            "type": "string"
          },
          "category": {
            "type": "object"
          },
          "subject": {
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "agency_responded",
              "escalated",
              "resolved",
              "closed"
            ],
            "type": "string"
          },
          "decision": {
            "type": "string"
          },
          "escalatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "closedAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "escalationAvailableAt": {
            "format": "date-time",
            "type": "string"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DisputeMessageShape"
            }
          }
        },
        "required": [
          "id",
          "bookingId",
          "packageTitle",
          "agencyId",
          "agencyName",
          "pilgrimName",
          "category",
          "subject",
          "status",
          "createdAt",
          "updatedAt",
          "escalationAvailableAt"
        ]
      },
      "DisputeMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        },
        "required": [
          "content"
        ]
      },
      "DisputeDecisionDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "minLength": 10,
            "maxLength": 2000
          }
        },
        "required": [
          "decision"
        ]
      },
...
    }
  }
}